New in 0.8.0
Fixed a bug where if a HTTP method wasn't specified, the previously
declared method was used, resulting in unexpected and unwanted results.

The second parameter in the `register` method is now optional. If not
set, all routes will be used by default. The exception to this is when
overriding the default route set for a single instance. In this case,
the asterisk ('*') may still be used.

QuickRoutes::register('users'); // assumes all
QuickRoutes::register('settings', '*', $settings_routes);

Introduced `with` method. Using `with` merges default routes with a set
of routes given for a single instance.

For example, this:

$default_routes = [
    'index' => [
        'pattern' => '/',
    ],
    'create' => [
        'pattern' => 'create',
        'methods' => ['get','post']
    ]
];

$image_routes = [
    'img_create' => [
        'pattern' => 'create/image'
        'methods' => ['post']
    ]
];

QuickRoutes::register('users', '*', array_merge($default_routes,
$image_routes));

is equivalent to this:

QuickRoutes::with($image_routes)->register('users');